+++ web/src/components/UI/index.js
@@ -1,3 +1,4 @@
 export { default as Header } from './Header';
 export { default as Footer } from './Footer';
 export { default as Layout } from './Layout';
+export { default as SpinLoader } from './SpinLoader';
