+++ web/src/containers/Home.js
@@ -2,16 +2,11 @@ import React, { Component } from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
-import { Button, Loader } from 'semantic-ui-react';
-import styled from '@emotion/styled';
+import { Button } from 'semantic-ui-react';
 import { TEST_IDS } from '../constants/ui';
 import { RequestList } from '../components/Request/RequestList';
+import { SpinLoader } from '../components/UI';
 import { getRequestsAction } from '../actionCreators';
-
-const StyledLoader = styled(Loader)`
-  font-size: 1rem;
-  padding: 0;
-`;
 
 export class Home extends Component {
   static displayName = '[Component Home]';
@@ -24,15 +19,7 @@ export class Home extends Component {
     const { requests, getRequestsStarted, errorMessage } = this.props;
     let requestsList = null;
     if (requests === null || getRequestsStarted) {
-      requestsList = (
-        <StyledLoader
-          active
-          size="small"
-          inline="centered"
-          content="Loading requests..."
-          data-testid={TEST_IDS.APP.LOADER}
-        />
-      );
+      requestsList = <SpinLoader text="Loading requests..." />;
     } else if (requests.length === 0) {
       requestsList = (
         <p data-testid={TEST_IDS.APP.EMPTY}>
